Refactor: Axios 객체들 axios.ts로 분리 및 try catch 내부에 toast alert 비즈니스 로직 구성하는 방식으로 수정
