+++ chatting-app/apis/roomApis.ts
@@ -2,37 +2,58 @@ import axios, { AxiosError } from "axios";
 import { SERVER_STATUS } from "../utils/enums";
 import { IRoom } from "../types/types";
 import { IRoomMoved } from "../utils/interfaces";
-import { requestWithTokenAxios } from "../utils/utils";
+import { requestWithTokenAxios } from "../utils/axios";
+import { handleAccessTokenErrors } from "../utils/utils";
+import { toast } from "react-toastify";
+import { toastConfig } from "../constants/etc";
 
 export const fetchAllRoomsList = async () => {
   let rooms = null;
   try {
-    const { data } = await axios.get<IRoom[]>(`/room/list`);
+    const { data } = await axios.get<IRoom[]>("/room/list");
     rooms = data;
   } catch (e) {
     const { response } = e as AxiosError;
-    console.log(`Error. Server status: ${response?.status}.`);
+    console.log(
+      `Failed to fetch room list. Server status: ${response?.status}.`
+    );
   }
   return rooms;
 };
 
+export const requestChangeToNewSubject = async (roomMovedInfo: IRoomMoved) => {
+  let isInValidToken = false;
+  try {
+    const { status } = await requestWithTokenAxios.put(
+      "/room/change_subject",
+      roomMovedInfo
+    );
+    if (status !== SERVER_STATUS.OK) throw new AxiosError(String(status));
+  } catch (e) {
+    const { response, message } = e as AxiosError;
+    const status = response?.status || +message;
+    if (status === SERVER_STATUS.UNAUTHORIZED) isInValidToken = true;
+    handleAccessTokenErrors(status);
+  }
+  return isInValidToken;
+};
+
 export const requestRoomDelete = async (targetRoomId: number) => {
   let isRoomDeleteSuccessful = false;
-  let isInvalidToken = false;
+  let isInValidToken = false;
   try {
     const { status } = await requestWithTokenAxios.delete(
       `/room/delete/${targetRoomId}`
     );
     if (status === SERVER_STATUS.OK) isRoomDeleteSuccessful = true;
-    else {
-      if (status !== SERVER_STATUS.FORBIDDEN) isInvalidToken = true;
-      throw new AxiosError(String(status));
-    }
+    else throw new AxiosError(String(status));
   } catch (e) {
-    const { message: status } = e as AxiosError;
-    // show toast;
+    const { response, message } = e as AxiosError;
+    const status = response?.status || +message;
+    if (status === SERVER_STATUS.UNAUTHORIZED) isInValidToken = true;
+    handleAccessTokenErrors(status);
   }
-  return [isRoomDeleteSuccessful, isInvalidToken];
+  return [isRoomDeleteSuccessful, isInValidToken];
 };
 
 export const requestToggleSubjectLike = async (
@@ -42,32 +63,16 @@ export const requestToggleSubjectLike = async (
   let isToggleSubjectLikeSuccessful = false;
   try {
     const { status } = await requestWithTokenAxios.post(
-      `/user/manage_subject_like`,
+      "/user/manage_subject_like",
       { subject, isAddLike }
     );
     if (status === SERVER_STATUS.OK) isToggleSubjectLikeSuccessful = true;
   } catch (e) {
-    const { message: status } = e as AxiosError;
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return isToggleSubjectLikeSuccessful;
-};
-
-export const requestChangeToNewSubject = async (roomMovedInfo: IRoomMoved) => {
-  let isInvalidToken = false;
-  try {
-    const { status } = await requestWithTokenAxios.put(
-      `/room/change_subject`,
-      roomMovedInfo
-    );
-    if (status !== SERVER_STATUS.FORBIDDEN) {
-      isInvalidToken = true;
-      throw new AxiosError(String(status));
-    }
-  } catch (e) {
-    // show toast;
-  }
-  return isInvalidToken;
 };
 
 export const fetchRoomsByKeyword = async (keyword: string) => {
@@ -76,9 +81,10 @@ export const fetchRoomsByKeyword = async (keyword: string) => {
     const { data } = await axios.get<Array<IRoom>>(
       `/room/search?keyword=${keyword}`
     );
+    if (!data || data.length === 0) throw new Error();
     searchedRooms = data;
   } catch (e) {
-    // show toast;
+    toast.error("No rooms have been found. :(", toastConfig);
   }
   return searchedRooms;
 };
@@ -89,13 +95,14 @@ export const checkIfIsPasswordCorrect = async (
 ) => {
   let isPasswordCorrect = false;
   try {
-    const { data } = await axios.post(`/room/enter_password`, {
+    const { data } = await axios.post<boolean>("/room/enter_password", {
       roomId,
       password,
     });
+    if (!data) throw new Error();
     isPasswordCorrect = data;
   } catch (e) {
-    // show toast;
+    toast.error("Password is not correct.", toastConfig);
   }
   return isPasswordCorrect;
 };
