+++ chatting-app/apis/userApis.ts
@@ -1,4 +1,4 @@
-import axios from "axios";
+import axios, { AxiosError } from "axios";
 import { SERVER_STATUS } from "../utils/enums";
 import { Iipdata, IParticipants } from "../types/types";
 import { IBannedUserList, ISignedIn } from "../utils/interfaces";
@@ -6,18 +6,23 @@ import {
   requestWithTokenAxios,
   signinAxios,
   signupAxios,
-} from "../utils/utils";
+} from "../utils/axios";
 import { API_KEY_REQUEST_URL } from "../constants/routes";
+import { handleAccessTokenErrors, handleNormalErrors } from "../utils/utils";
+import { toastConfig } from "../constants/etc";
+import { toast } from "react-toastify";
 
 export const fetchUserInfo = async () => {
   let userInfo = null;
   try {
     const { data } = await requestWithTokenAxios.get<ISignedIn>(
-      `/user/get-userInfo`
+      "/user/get-userInfo"
     );
     if (data) userInfo = data;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return userInfo;
 };
@@ -25,10 +30,12 @@ export const fetchUserInfo = async () => {
 export const requestSignOut = async () => {
   let isSignOutSuccessful = false;
   try {
-    const { status } = await requestWithTokenAxios.get(`/user/signout`);
+    const { status } = await requestWithTokenAxios.get("/user/signout");
     if (status === SERVER_STATUS.OK) isSignOutSuccessful = true;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return isSignOutSuccessful;
 };
@@ -43,7 +50,8 @@ export const fetchUserPrivateIpAddress = async () => {
     );
     userPrivateIpAddress = ip;
   } catch (e) {
-    // show toast;
+    console.log(e);
+    if (!userPrivateIpAddress) throw new Error();
   }
   return userPrivateIpAddress;
 };
@@ -56,7 +64,9 @@ export const fetchRoomParticipants = async (roomId: number) => {
     );
     participants = data;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleNormalErrors(status);
   }
   return participants;
 };
@@ -69,7 +79,9 @@ export const fetchBannedUserList = async (roomId: number) => {
     );
     bannedUserList = data;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return bannedUserList;
 };
@@ -77,21 +89,27 @@ export const fetchBannedUserList = async (roomId: number) => {
 export const requestUserExpel = async (
   roomId: number,
   ipAddress: string,
-  userName: string
+  userName: string,
+  isBanUser: boolean
 ) => {
   let isUserBanSuccessful = false;
   try {
-    const { status } = await axios.post(`/user/add_banned`, {
+    const { status } = await axios.post("/user/add_banned", {
       roomId,
       ipAddress,
       userName,
     });
     if (status === SERVER_STATUS.OK) {
-      // show toast; -> "This room is disbanded." or "You are banned!"
+      const banMessage = isBanUser
+        ? "You are banned!"
+        : "This room is disbanded.";
+      toast.error(banMessage, toastConfig);
       isUserBanSuccessful = true;
     }
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleNormalErrors(status);
   }
   return isUserBanSuccessful;
 };
@@ -102,13 +120,15 @@ export const unlockBannedIpAddress = async (
 ) => {
   let isUnlockSuccessful = false;
   try {
-    const { status } = await requestWithTokenAxios.post(`/room/unlock_ban`, {
+    const { status } = await requestWithTokenAxios.post("/room/unlock_ban", {
       bannedIpNo,
       roomId,
     });
     if (status === SERVER_STATUS.OK) isUnlockSuccessful = true;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return isUnlockSuccessful;
 };
@@ -119,7 +139,9 @@ export const fetchUserSettingsInfo = async () => {
     const { data } = await requestWithTokenAxios.get(`/user/info`);
     userSettingsInfo = data;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return userSettingsInfo;
 };
@@ -128,11 +150,13 @@ export const requestAlterUserSettingsInfo = async (formData: FormData) => {
   let isAlterSuccessful = false;
   let isInvalidToken = false;
   try {
-    const { status } = await requestWithTokenAxios.put(`/user/alter`, formData);
+    const { status } = await requestWithTokenAxios.put("/user/alter", formData);
     if (status === SERVER_STATUS.OK) isAlterSuccessful = true;
   } catch (e) {
     isInvalidToken = true;
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return [isAlterSuccessful, isInvalidToken];
 };
@@ -141,13 +165,15 @@ export const requestWithdrawal = async (inputPassword: string) => {
   let isWithdrawalSuccessful = false;
   let isInvalidToken = false;
   try {
-    const { status } = await requestWithTokenAxios.put(`/user/withdraw`, {
+    const { status } = await requestWithTokenAxios.put("/user/withdraw", {
       inputPassword,
     });
     if (status === SERVER_STATUS.OK) isWithdrawalSuccessful = true;
   } catch (e) {
     isInvalidToken = true;
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return [isWithdrawalSuccessful, isInvalidToken];
 };
@@ -155,13 +181,15 @@ export const requestWithdrawal = async (inputPassword: string) => {
 export const requestSignIn = async (id: string, password: string) => {
   let userInfo = null;
   try {
-    const { data } = await signinAxios.post<ISignedIn>(`/user/signin`, {
+    const { data } = await signinAxios.post<ISignedIn>("/user/signin", {
       id,
       password,
     });
     userInfo = data;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleAccessTokenErrors(status);
   }
   return userInfo;
 };
@@ -169,12 +197,14 @@ export const requestSignIn = async (id: string, password: string) => {
 export const requestSignUp = async (formData: FormData) => {
   let isSignUpSuccessful = false;
   try {
-    const { status } = await signupAxios.post(`/user/signup`, formData, {
+    const { status } = await signupAxios.post("/user/signup", formData, {
       headers: { "Content-Type": "multipart/form-data" },
     });
     if (status === SERVER_STATUS.OK) isSignUpSuccessful = true;
   } catch (e) {
-    // show toast;
+    const { response } = e as AxiosError;
+    const status = response?.status;
+    if (status) handleNormalErrors(status);
   }
   return isSignUpSuccessful;
 };
